Move Script Inspector, add Saving Shortcut, and Saving for  Shader
