+++ saves/render_render.toml
@@ -1,0 +1,20 @@
+id = "render"
+name = "render"
+
+[transform.position]
+x = 0.0
+y = 0.0
+z = 0.0
+
+[transform.rotation]
+h = 0.0
+p = 0.0
+r = 0.0
+
+[transform.scale]
+x = 1.0
+y = 1.0
+z = 1.0
+
+[properties.script_properties."C:/Users/user/Desktop/RaytopianProjects/code/p3deditor2/PANDA3D-EDITOR/script_inspector_example.py"]
+texture = "file:///C:/Users/user/Desktop/RaytopianProjects/code/p3deditor2/PANDA3D-EDITOR/Temp_Brush.png"
